Fix cast between integer and pointer of different size

+++ src/Cedar/Command.c
@@ -408,7 +408,7 @@ void CheckThread1(THREAD *thread, void *param)
 	{
 		Lock(ct1->lock);
 		check_thread_global_1 = ct1->num;
-		InputToNull((void *)check_thread_global_1);
+		InputToNull((void *)(intptr_t)check_thread_global_1);
 		check_thread_global_1 = check_thread_global_1 + 1 + RetZero();
 		ct1->num = check_thread_global_1;
 		Unlock(ct1->lock);
@@ -695,7 +695,7 @@ bool CheckMemory()
 		pp[i] = ReAlloc(pp[i], size);
 		for (j = old_size;j < size;j++)
 		{
-			InputToNull((void *)(UINT)(((UCHAR *)pp[i])[j] = j % 256));
+			InputToNull((void *)(intptr_t)(((UCHAR *)pp[i])[j] = j % 256));
 		}
 	}
 	for (i = 0;i < num;i++)
@@ -7324,13 +7324,13 @@ LIST *StrToPortList(char *str)
 			FreeToken(t);
 			return NULL;
 		}
-		if (IsInList(o, (void *)n))
+		if (IsInList(o, (void *)(intptr_t)n))
 		{
 			ReleaseList(o);
 			FreeToken(t);
 			return NULL;
 		}
-		Add(o, (void *)n);
+		Add(o, (void *)(intptr_t)n);
 	}
 
 	FreeToken(t);
@@ -7407,7 +7407,7 @@ UINT PsClusterSettingMember(CONSOLE *c, char *cmd_name, wchar_t *str, void *para
 
 		for (i = 0;i < t.NumPort;i++)
 		{
-			t.Ports[i] = (UINT)LIST_DATA(ports, i);
+			t.Ports[i] = (intptr_t)LIST_DATA(ports, i);
 		}
 
 		t.Weight = weight;
@@ -7479,7 +7479,7 @@ bool CmdEvalHostAndPort(CONSOLE *c, wchar_t *str, void *param)
 
 	tmp = CopyUniToStr(str);
 
-	ret = ParseHostPort(tmp, NULL, NULL, (UINT)param);
+	ret = ParseHostPort(tmp, NULL, NULL, (intptr_t)param);
 
 	if (ret == false)
 	{
